Added construction functions.
These will become part of a new extension. It allows for primitives to
be defined by other primitives. Eg arc by three vecs and a call to
Arc.prototype.fromVec3 that fits the arc to 3 points. Constructions are
defined by the API user, and can be nested and will provied a easy way
to construct very complex primitive constructions